Extract repositories to directory
