+++ game/src/game.js
@@ -6,7 +6,7 @@ import EventEmitter from 'eventemitter3'
 import { Event, Channel } from 'common'
 import { GAME_WIDTH, GAME_HEIGHT } from './constant/rendering'
 import { State, state } from './state'
-import playerRepository from './playerRepository'
+import playerRepository from './repository/player'
 import { transitionToRoundEnd } from './roundEnd'
 import Layer from './constant/layer'
 import countdown from './countdown'
