+++ game/src/index.js
@@ -12,7 +12,7 @@ import Scene from './Scene'
 import Layer from './constant/layer'
 import fullscreenFadeInOut from './fullscreenFadeInOut'
 import { State, state } from './state'
-import playerRepository from './playerRepository'
+import playerRepository from './repository/player'
 import { GAME_WIDTH, GAME_HEIGHT } from './constant/rendering'
 import GameEvent from './constant/gameEvent'
 import * as qrCode from './qrCode'
