+++ game/test/repository/index.test.js
@@ -1,5 +1,5 @@
-import { state } from '../src/state'
-import repo from '../src/repository'
+import { state } from '../../src/state'
+import repo from '../../src/repository'
 
 beforeEach(() => {
   state.players = [
